requirements.txt
python-dotenv
requests
pandas
numpy
matplotlib
scikit-learn
streamlit
yfinance
